annotator/requirements.txt
Pillow
opencv-python
face_recognition
redis
numpy